requirements.txt
pdfplumber
spacy
pymongo
mysql-connector-python
python-dotenv
nbconvert
papermill
jupyter
flask
gunicorn